requirements.txt
python-dotenv==0.21.0
aiogram==2.22.2
asyncpg==0.26.0
asyncio==3.4.3